--- public/usage-examples/geometry/circle_at-1-example-oop.cs
@@ -1,17 +1,0 @@
-﻿using SplashKitSDK;
-
-public class Program
-{
-    public static void Main()
-    {
-        Window window = new Window("Circle At", 800, 600);
-
-        Circle circle = SplashKit.CircleAt(200, 300, 100);
-
-        window.Clear(Color.White);
-        SplashKit.FillCircle(Color.Red, circle);
-        window.Refresh();
-
-        SplashKit.Delay(8000);
-    }
-}
