--- public/usage-examples/geometry/circle_at-1-example.txt
@@ -1,1 +1,0 @@
-Using Circle At
