+++ public/usage-examples/geometry/circle_at_from_points-1-example.png
